Basics of zoom implemented

+++ prey.cpp
@@ -37,7 +37,7 @@ void Prey::advance(int phase)
 {
     if (!phase || (vx == 0 && vy == 0) || isDied) return;
 
-    int unit = dynamic_cast<MyQGraphicsView*>(scene()->parent())->getUnit();
+    double unit = dynamic_cast<MyQGraphicsView*>(scene()->parent())->getUnit();
     moveBy(vx*unit/100., -vy*unit/100.); // Timer shots 100 p/s, so movement velocity is V*unit pixels per second
     curr += QPointF(vx/100., vy/100.);
