+++ yerp.cpp
@@ -35,7 +35,7 @@ void Yerp::advance(int phase)
 {
     if (!phase || (vx == 0 && vy == 0)) return;
 
-    int unit = dynamic_cast<MyQGraphicsView*>(scene()->parent())->getUnit();
+    double unit = dynamic_cast<MyQGraphicsView*>(scene()->parent())->getUnit();
     moveBy(vx*unit/100., -vy*unit/100.); // Timer shots 100 p/s, so movement velocity is V*unit pixels per second
     curr += QPointF(vx/100., vy/100.);
 }
